fix: handle race conditions accessing elements in DOM
Fixed the browser to not appear to be 'ready' to set the body before the
text to be set has been processed.

Otherwise, because the content is set asynchronously in a browser after
the call to browser.setText, it can be that the first call to
document.readyState is done before the browser has processed the request
to modify the page, and therefore document.readyState returns true even
if the page has no content. To fix this, we check that the document has
at least one element before asking for document.readyState.