fix: handle race conditions accessing elements in DOM
Fixed the browser to not appear to be 'ready' to set the body before the
text is set. Otherwise, because the content is set asynchronously in a
browser, there is a race between setting the text and writing the body.

Also, check for the document not being null before accessing some of its
properties.